convert BspServerRunner.runSession into MillMain.runBspSession
